4-gpus-pad-right

+++ lit_llama/model.py
@@ -40,7 +40,7 @@ def apply_rope(x: torch.Tensor, rope_cache: torch.Tensor) -> torch.Tensor:
     # FIXME remove this comment
     # Converting to complex here because complex buffers not supported in PyTorch
     # RuntimeError: Input tensor data type is not supported for NCCL process group: ComplexFloat
-    rope_cache = torch.polar(torch.ones_like(rope_cache[:T]), rope_cache[:T])  # complex64
+    rope_cache = torch.polar(torch.ones_like(rope_cache[:T], dtype=torch.float), rope_cache[:T].float())  # complex64
 
     # cast because `view_as_complex` does not support 16 bit tensors
     xc = torch.view_as_complex(x.float().reshape(*x.shape[:-1], -1, 2))
